Some minor metadata fixes
Rename of some obj files as txt for Android testing

Sample okay on Android, tvos, Mac... mixed results on iOS

Some camera controls for same

+++ engine_testing/custom-objects-test4/shared/DepthEnvironment.cpp
@@ -54,7 +54,7 @@ DepthEnvironment * InitDepthEnvironment( lua_State * L )
         #else
             glClearDepth
         #endif
-            ( GLdouble( clear ) ); // TODO: is this expensive? if so, avoid when possible...
+            ( clear ); // TODO: is this expensive? if so, avoid when possible...
             glClear( GL_DEPTH_BUFFER_BIT );
 		};
